+++ src/cpcrypto/symmetric/sha1.h
@@ -54,7 +54,7 @@ class Sha1 {
 #endif
   void Init();
   bool Transform(u32*);
-  void Update(const byte*, const u32 long);
+  void Update(const byte*, const u32);
   void Final();
   void getDigest(byte* out);
 };
